Write fizzbuzz code and readme file
Wrote Readme file which contains the problem solving exercise plan and
pseudocode. Wrote corresponding fizzbuzz code and ensured it runs
as expected. Test ran succesfully other than for condition 3 fizzbuzz
was returning fizz instead of fizzbuzz for numbers divisible by 3 and 5.
Corrected this by moving the condition so that it ran first. Program
does what it's supposed to for lower numbers, however it began
freezing the browser with inputs greater than ~400,000.